Move 'reimburseFee' to separate module
